Creating a new class for config creation
