+++ genai-perf/tests/test_cli.py
@@ -34,6 +34,7 @@ from genai_perf import __version__, parser
 from genai_perf.config.generate.perf_analyzer_config import PerfAnalyzerConfig
 from genai_perf.config.input.config_command import ConfigCommand
 from genai_perf.config.input.config_defaults import EndPointDefaults
+from genai_perf.config.input.create_config import CreateConfig
 from genai_perf.inputs.input_constants import (
     AudioFormat,
     ImageFormat,
@@ -77,7 +78,8 @@ class TestCLIArguments:
         monkeypatch.setattr("sys.argv", ["genai-perf"] + args)
 
         with pytest.raises(SystemExit) as excinfo:
-            parser.parse_args()
+            args, _ = parser.parse_args()
+            config = CreateConfig.create(args)
 
         # Check that the exit was successful
         assert excinfo.value.code == 0
@@ -525,7 +527,8 @@ class TestCLIArguments:
         logging.init_logging()
         combined_args = self.base_args + arg
         monkeypatch.setattr("sys.argv", combined_args)
-        args, config, _ = parser.parse_args()
+        args, _ = parser.parse_args()
+        config = CreateConfig.create(args)
 
         # Check that the attributes are set correctly
         for key, value in expected_cli_attributes.items():
@@ -551,7 +554,8 @@ class TestCLIArguments:
             "fakefile.txt",
         ]
         monkeypatch.setattr("sys.argv", combined_args)
-        args, config, _ = parser.parse_args()
+        args, _ = parser.parse_args()
+        config = CreateConfig.create(args)
         assert args.input_file == Path(
             "fakefile.txt"
         ), "The file argument should be the path to the file"
@@ -603,9 +607,9 @@ class TestCLIArguments:
         logging.init_logging()
         combined_args = self.base_args + arg
         monkeypatch.setattr("sys.argv", combined_args)
-        args, _, _ = parser.parse_args()
+        args, _ = parser.parse_args()
         config = ConfigCommand({"model_name": args.model})
-        config = parser.add_cli_options_to_config(config, args)
+        config = CreateConfig._add_cli_options_to_config(config, args)
         config.infer_and_check_options()
         perf_analyzer_config = PerfAnalyzerConfig(config)
 
@@ -645,10 +649,10 @@ class TestCLIArguments:
         logging.init_logging()
         combined_args = self.base_args + arg
         monkeypatch.setattr("sys.argv", combined_args)
-        args, _, _ = parser.parse_args()
+        args, _ = parser.parse_args()
 
         config = ConfigCommand({"model_names": args.model})
-        config = parser.add_cli_options_to_config(config, args)
+        config = CreateConfig._add_cli_options_to_config(config, args)
         config.infer_and_check_options()
         perf_analyzer_config = PerfAnalyzerConfig(config)
 
@@ -657,7 +661,8 @@ class TestCLIArguments:
     def test_default_load_level(self, monkeypatch, capsys):
         logging.init_logging()
         monkeypatch.setattr("sys.argv", self.base_args)
-        _, config, _ = parser.parse_args()
+        args, _ = parser.parse_args()
+        config = CreateConfig.create(args)
         assert config.perf_analyzer.stimulus["concurrency"] == 1
 
     def test_load_manager_args_with_payload(self, monkeypatch, mocker):
@@ -672,7 +677,8 @@ class TestCLIArguments:
             ],
         )
         mocker.patch.object(Path, "is_file", return_value=True)
-        args, config, _ = parser.parse_args()
+        args, _ = parser.parse_args()
+        config = CreateConfig.create(args)
         assert args.concurrency is None
         assert config.perf_analyzer.get_field("stimulus").is_set_by_user is False
 
@@ -686,7 +692,8 @@ class TestCLIArguments:
         )
 
         with pytest.raises(SystemExit) as excinfo:
-            parser.parse_args()
+            args, _ = parser.parse_args()
+            config = CreateConfig.create(args)
 
         assert excinfo.value.code != 0
         captured = capsys.readouterr()
@@ -697,14 +704,15 @@ class TestCLIArguments:
         expected_error_message = "Required field model_names is not set"
 
         with pytest.raises(ValueError) as execinfo:
-            parser.parse_args()
+            args, _ = parser.parse_args()
+            config = CreateConfig.create(args)
 
         assert expected_error_message == execinfo.value.args[0]
 
     def test_pass_through_args(self, monkeypatch):
         other_args = ["--", "With", "great", "power"]
         monkeypatch.setattr("sys.argv", self.base_args + other_args)
-        _, _, pass_through_args = parser.parse_args()
+        _, pass_through_args = parser.parse_args()
 
         assert pass_through_args == other_args[1:]
 
@@ -721,7 +729,8 @@ class TestCLIArguments:
         )
 
         with pytest.raises(ValueError) as execinfo:
-            parser.parse_args()
+            args, _ = parser.parse_args()
+            config = CreateConfig.create(args)
 
         expected_error_message = "Unknown argument passed to GenAI-Perf: --wrong-arg"
         assert expected_error_message == execinfo.value.args[0]
@@ -732,7 +741,8 @@ class TestCLIArguments:
             ["genai-perf", "create-template", "custom_template.yaml"],
         )
 
-        args, config, _ = parser.parse_args()
+        args, _ = parser.parse_args()
+        config = CreateConfig.create(args)
         assert config.template_filename == Path("custom_template.yaml")
 
     @pytest.mark.parametrize(
@@ -996,7 +1006,8 @@ class TestCLIArguments:
         monkeypatch.setattr("sys.argv", args)
 
         with pytest.raises(ValueError) as execinfo:
-            parser.parse_args()
+            args, _ = parser.parse_args()
+            config = CreateConfig.create(args)
 
         assert expected_error_message == execinfo.value.args[0]
 
@@ -1041,7 +1052,8 @@ class TestCLIArguments:
     def test_inferred_output_format(self, monkeypatch, args, expected_format):
         monkeypatch.setattr("sys.argv", self.base_args + args)
 
-        _, config, _ = parser.parse_args()
+        args, _ = parser.parse_args()
+        config = CreateConfig.create(args)
         assert config.endpoint.output_format == expected_format
 
     @pytest.mark.parametrize(
@@ -1075,7 +1087,8 @@ class TestCLIArguments:
         monkeypatch.setattr("sys.argv", combined_args)
 
         with pytest.raises(ValueError) as execinfo:
-            parser.parse_args()
+            args, _ = parser.parse_args()
+            config = CreateConfig.create(args)
 
         assert expected_error_message == execinfo.value.args[0]
 
@@ -1093,7 +1106,8 @@ class TestCLIArguments:
         monkeypatch.setattr("sys.argv", combined_args)
 
         with pytest.raises(ValueError) as execinfo:
-            parser.parse_args()
+            args, _ = parser.parse_args()
+            config = CreateConfig.create(args)
 
         assert expected_error_message == execinfo.value.args[0]
 
@@ -1130,7 +1144,8 @@ class TestCLIArguments:
         mocker.patch.object(Path, "is_file", return_value=True)
         combined_args = self.base_args + args
         monkeypatch.setattr("sys.argv", combined_args)
-        parsed_args, config, _ = parser.parse_args()
+        parsed_args, _ = parser.parse_args()
+        config = CreateConfig.create(parsed_args)
         assert config.input.prompt_source == expected_prompt_source
         assert config.input.payload_file == expected_input_file
 
@@ -1151,7 +1166,8 @@ class TestCLIArguments:
         combined_args = self.base_args + args
         monkeypatch.setattr("sys.argv", combined_args)
         with pytest.raises(ValueError):
-            parser.parse_args()
+            args, _ = parser.parse_args()
+            config = CreateConfig.create(args)
 
     def test_inferred_prompt_source_invalid_input(self, monkeypatch, mocker):
         arg = ["--input-file", "invalid_input"]
@@ -1160,7 +1176,8 @@ class TestCLIArguments:
         combined_args = self.base_args + arg
         monkeypatch.setattr("sys.argv", combined_args)
         with pytest.raises(SystemExit):
-            parser.parse_args()
+            args, _ = parser.parse_args()
+            config = CreateConfig.create(args)
 
     @pytest.mark.parametrize(
         "args",
@@ -1177,7 +1194,8 @@ class TestCLIArguments:
         monkeypatch.setattr("sys.argv", combined_args)
 
         with pytest.raises(ValueError) as excinfo:
-            parser.parse_args()
+            args, _ = parser.parse_args()
+            config = CreateConfig.create(args)
 
     @pytest.mark.parametrize(
         "args",
@@ -1199,7 +1217,8 @@ class TestCLIArguments:
         monkeypatch.setattr("sys.argv", combined_args)
 
         with pytest.raises(ValueError) as excinfo:
-            parser.parse_args()
+            args, _ = parser.parse_args()
+            config = CreateConfig.create(args)
 
     @pytest.mark.parametrize(
         "args , expected_error_message",
@@ -1237,7 +1256,8 @@ class TestCLIArguments:
         mocker.patch.object(Path, "is_file", return_value=True)
         monkeypatch.setattr("sys.argv", combined_args)
         with pytest.raises(ValueError) as execinfo:
-            parser.parse_args()
+            args, _ = parser.parse_args()
+            config = CreateConfig.create(args)
 
         assert expected_error_message == execinfo.value.args[0]
 
@@ -1249,7 +1269,8 @@ class TestCLIArguments:
         mocker.patch.object(Path, "is_file", return_value=True)
         monkeypatch.setattr("sys.argv", valid_args)
         try:
-            parser.parse_args()
+            args, _ = parser.parse_args()
+            config = CreateConfig.create(args)
         except SystemExit:
             pytest.fail("Unexpected error in test")
 
@@ -1269,11 +1290,12 @@ class TestCLIArguments:
             "100",
         ]
         logging.init_logging()
-        logger = logging.getLogger("genai_perf.parser")
+        logger = logging.getLogger("genai_perf.config.input.create_config")
         mocker.patch.object(Path, "is_file", return_value=True)
         monkeypatch.setattr("sys.argv", args)
         with patch.object(logger, "warning") as mock_logger:
-            parser.parse_args()
+            args, _ = parser.parse_args()
+            config = CreateConfig.create(args)
         mock_logger.assert_any_call(expected_warning_message)
 
     @pytest.mark.parametrize(
@@ -1354,7 +1376,8 @@ class TestCLIArguments:
     )
     def test_server_metrics_url_arg_valid(self, args_list, expected_url, monkeypatch):
         monkeypatch.setattr("sys.argv", args_list)
-        args, config, _ = parser.parse_args()
+        args, _ = parser.parse_args()
+        config = CreateConfig.create(args)
 
         if expected_url:
             assert config.endpoint.server_metrics_urls == expected_url
@@ -1377,7 +1400,8 @@ class TestCLIArguments:
             "test_revision",
         ]
         monkeypatch.setattr("sys.argv", args)
-        parsed_args, config, _ = parser.parse_args()
+        parsed_args, _ = parser.parse_args()
+        config = CreateConfig.create(parsed_args)
 
         assert parsed_args.tokenizer == "test_tokenizer"
         assert parsed_args.tokenizer_trust_remote_code
@@ -1397,7 +1421,8 @@ class TestCLIArguments:
         monkeypatch.setattr("sys.argv", combined_args)
 
         with pytest.raises(SystemExit) as excinfo:
-            parser.parse_args()
+            args, _ = parser.parse_args()
+            config = CreateConfig.create(args)
 
         assert excinfo.value.code != 0
         captured = capsys.readouterr()
@@ -1418,7 +1443,8 @@ class TestCLIArguments:
 
         monkeypatch.setattr("sys.argv", combined_args)
         with pytest.raises(ValueError) as execinfo:
-            parser.parse_args()
+            args, _ = parser.parse_args()
+            config = CreateConfig.create(args)
 
         expected_error_message = "In order to use the CLI to override the config, the --override-config flag must be set."
         assert expected_error_message == execinfo.value.args[0]
@@ -1437,7 +1463,8 @@ class TestCLIArguments:
         ]
 
         monkeypatch.setattr("sys.argv", combined_args)
-        _, config, _ = parser.parse_args()
+        args, _ = parser.parse_args()
+        config = CreateConfig.create(args)
 
         assert config.model_names == ["test_model_name"]
         assert config.perf_analyzer.stimulus == {"request_rate": 100}
@@ -1456,6 +1483,7 @@ class TestCLIArguments:
         ]
 
         monkeypatch.setattr("sys.argv", combined_args)
-        _, config, _ = parser.parse_args()
+        args, _ = parser.parse_args()
+        config = CreateConfig.create(args)
 
         assert config.verbose is True
